requirements.txt
chromadb
langchain
langchain-community
azure-ai-documentintelligence
streamlit
pydantic
grpcio
shortuuid
openai
open-clip-torch
fastapi
uvicorn
aiofiles
azure-ai-formrecognizer
python-multipart
sqlalchemy
watchdog
ipython
tzlocal
pillow
matplotlib